Move output(from:) to PipeOutputProtocol and rename files to better suited as a package
